+++ design/components/breadcrumb/breadcrumb.component.ts
@@ -1,0 +1,35 @@
+import {
+  Component,
+  component,
+  applyTemplate,
+  applyStyles,
+} from "@apexdesigner/dsl/component";
+import { BreadcrumbLevelComponent } from "@components";
+
+/**
+ * Breadcrumb
+ *
+ * Navigation breadcrumb container that displays child breadcrumb levels separated by delimiters.
+ */
+@component({ allowChildren: true })
+export class BreadcrumbComponent extends Component {
+  levels!: BreadcrumbLevelComponent[];
+}
+
+applyTemplate(
+  BreadcrumbComponent,
+  `
+  <ng-content></ng-content>
+`,
+);
+
+applyStyles(
+  BreadcrumbComponent,
+  `
+  :host {
+    display: flex;
+    align-items: center;
+    font-size: 14px;
+  }
+`,
+);