+++ design/project.ts
@@ -5,6 +5,10 @@ export class BaseLibrary extends Project {
 
   isLibrary = true;
 
+  designDependencies = [
+    { package: "@apexdesigner/doc-generators", versionSelector: '^0.0.8' },
+  ];
+
   parameterValues = {
     formFieldAppearance: "outline",
     formFieldSubscriptSizing: "dynamic",